Feat: #458[유영] Redis 구조 전체 변경

See merge request s09-webmobile1-sub2/S09P12A602!95

+++ backend/oho/src/main/java/com/ssafy/oho/model/service/PlayerService.java
@@ -11,14 +11,10 @@ import com.ssafy.oho.util.exception.PlayerGetException;
 import com.ssafy.oho.util.exception.PlayerSetException;
 import com.ssafy.oho.util.exception.PlayerUpdateException;
 import io.openvidu.java.client.*;
-import org.springframework.data.redis.core.HashOperations;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Random;
+import java.util.*;
 
 @Service
 public class PlayerService extends RedisService {
@@ -38,7 +34,7 @@ public class PlayerService extends RedisService {
     public PlayerResponseDto setHead(PlayerRequestDto playerRequestDto, String roomId, OpenVidu openVidu) throws PlayerSetException {
         try {
             PlayerResponseDto headResponseDto = setPlayer(playerRequestDto, roomId, openVidu);
-            Player head = playerRepository.findById(headResponseDto.getId()).orElseThrow(()-> new PlayerSetException());
+            Player head = playerRepository.findById(headResponseDto.getId()).orElseThrow(PlayerSetException::new);
 
             /*** Entity Build ***/
             head = Player.builder()
@@ -51,39 +47,39 @@ public class PlayerService extends RedisService {
             playerRepository.save(head);
 
             /*** Redis Input ***/
-            super.hashOperations.put(super.getPlayerListKey(roomId, head.getId()), "head", Boolean.toString(head.isHead()));
-            super.hashOperations.put(super.getPlayerListKey(roomId, head.getId()), "ready", "true");
+            Map<String, String> hash = new HashMap<>() {{ put("head", "true"); put("ready", "true"); }};
+            super.setPlayerInfo(roomId, head.getId(), hash);
 
             /*** Response DTO Build ***/
             headResponseDto = PlayerResponseDto.builder()
                     .id(head.getId())
                     .nickname(head.getNickname())
                     .head(head.isHead())
+                    .ready(true)
                     .build();
 
             return headResponseDto;
         } catch (Exception e) {
+            e.printStackTrace();
             throw new PlayerSetException();
         }
     }
     public PlayerResponseDto setPlayer(PlayerRequestDto playerRequestDto, String roomId, OpenVidu openVidu) throws PlayerSetException {
         try {
-            Room room = roomRepository.findById(roomId).orElseThrow(()-> new PlayerSetException());
-            // 방이 존재하지 않을 경우
-            // 플레이어가 4명 이상인 경우
-            // 게임이 이미 시작된 경우
-            if (room == null || 4 <= playerRepository.countByRoom(room) || room.isProgress()) {
+            Room room = roomRepository.findById(roomId).orElseThrow(PlayerSetException::new);
+            // 방이 존재하지 않을 경우, 플레이어가 4명 이상인 경우, 게임이 이미 시작된 경우
+            if (room == null ||(room.getPlayers() != null && 4 <= room.getPlayers().size()) || room.isProgress()) {
                 throw new PlayerSetException("해당 방에 접속할 수 없습니다.");
             }
 
             // 닉네임 입력 시 해당 닉네임, 미입력시 랜덤 닉네임 생성
             String nickname = "";
-            if (playerRequestDto.getNickname() == null && playerRequestDto.getNickname().trim().equals("")) {  // 닉네임 없을 경우
+            if (playerRequestDto.getNickname() == null || playerRequestDto.getNickname().trim().equals("")) {  // 닉네임 없을 경우
                 nickname = playerRequestDto.getNickname();
             } else {  // 닉네임 없을 경우
                 nickname = randAdj[new Random().nextInt(randAdj.length)] +
                         randNoun[new Random().nextInt(randNoun.length)] +
-                        new Random().nextInt(10000);  // 랜덤 닉네임 생성
+                        (new Random().nextInt(8999) + 1000);  // 랜덤 닉네임 생성 알고리즘
             }
 
             /* 혜지 : OpenVidu Token 발급 */
@@ -97,29 +93,30 @@ public class PlayerService extends RedisService {
                     .data(nickname) /* 혜지: data로 nickname 실어 보내기 */
                     .build();
             Connection connection = session.createConnection(properties);
-            String token=connection.getToken();  // VALUE EXAMPLE : "wss://localhost:4443?sessionId=ses_JM9v0nfD1l&token=tok_MIYGGzuDQb8Xf1Qd"
+            String token = connection.getToken();  // VALUE EXAMPLE : "wss://localhost:4443?sessionId=ses_JM9v0nfD1l&token=tok_MIYGGzuDQb8Xf1Qd"
+            String[] tokenUnit = token.split("=");
+            String tokenParam = tokenUnit[tokenUnit.length - 1];
 
             /*** Entity Build ***/
             Player player = Player.builder()
-                    .id(token)
+                    .id(tokenParam)
                     .room(room)
                     .nickname(nickname)
                     .head(false)
                     .build();
 
             playerRepository.save(player);
-            defaultPlayerRedis(roomId, player); // Redis에 저장
+            super.setPlayer(roomId, player);
 
             /*** Response DTO Build ***/
-            PlayerResponseDto playerResponseDto = PlayerResponseDto.builder()
+            return PlayerResponseDto.builder()
                     .id(player.getId())
                     .nickname(player.getNickname())
                     .head(player.isHead())
+                    .ready(Boolean.parseBoolean((String) super.getPlayerInfo(roomId, player.getId(), "ready")))
                     .build();
-
-            return playerResponseDto;
-        } catch (Exception e) {//OpenViduJavaClientException, OpenViduHttpException, ...
-            //e.printStackTrace();
+        } catch (Exception e) {  // OpenViduJavaClientException, OpenViduHttpException, ...
+            e.printStackTrace();
             throw new PlayerSetException();
         }
     }
@@ -129,13 +126,12 @@ public class PlayerService extends RedisService {
             Player player = playerRepository.findById(playerRequestDto.getId()).orElseThrow(()-> new PlayerGetException());
 
             /*** Response DTO Build ***/
-            PlayerResponseDto playerResponseDto = PlayerResponseDto.builder()
+            return PlayerResponseDto.builder()
                     .id(player.getId())
                     .nickname(player.getNickname())
                     .head(player.isHead())
+                    .ready((boolean) super.getPlayerInfo(playerRequestDto.getRoomId(), player.getId(), "ready"))
                     .build();
-
-            return playerResponseDto;
         } catch(Exception e) {
             //System.out.println(e.getMessage());
             throw new PlayerGetException();
@@ -147,13 +143,13 @@ public class PlayerService extends RedisService {
         try {
             /*** 유효성 검사 ***/
             // 현재 방의 플레이어 존재 확인
-            Player player = playerRepository.findById(playerRequestDto.getId()).orElseThrow(()-> new PlayerGetException());
+            Player player = playerRepository.findById(playerRequestDto.getId()).orElseThrow(PlayerGetException::new);
             if(player == null || !player.getRoom().getId().equals(roomId)){
                 throw new PlayerGetException();
             }
 
             // 방 유효성 확인
-            Room room = roomRepository.findById(roomId).orElseThrow(()-> new PlayerGetException());
+            Room room = roomRepository.findById(roomId).orElseThrow(PlayerGetException::new);
             if(room == null || room.getId().trim().equals("")){
                 throw new PlayerGetException();
             }
@@ -189,31 +185,31 @@ public class PlayerService extends RedisService {
 
             String playerId = (String) payload.get("playerId");
             // Redis 저장 확인
-            if (!hashOperations.hasKey(getPlayerListKey(roomId, playerId), "ready")) {
+            if (super.getPlayer(roomId, playerId) == null) {
                 // Redis에도 DB에도 없을 시 Exception
                 if(!playerRepository.existsById(playerId)) {
                     throw new PlayerUpdateException();
                 }
-                
+
                 // DB엔 있으나 Redis에 없을 시 Redis에 저장
-                defaultPlayerRedis(roomId, playerRepository.findById(playerId).orElseThrow(()-> new PlayerUpdateException()));
+                super.setPlayer(roomId, playerRepository.findById(playerId).orElseThrow(()-> new PlayerUpdateException()));
             }
 
             /*** Redis Input ***/
-            if(payload.containsKey("ready")) {  // Ready 상태 변경
-                super.hashOperations.put(getPlayerListKey(roomId, playerId), "ready", Boolean.toString((boolean) payload.get("ready")));
+            Map<String, String> hash = new HashMap<>();
+            if(payload.containsKey("head")) {  // Head 상태 변경
+                hash.put("head", (String) payload.get("head"));
             }
+            if(payload.containsKey("ready")) {  // Ready 상태 변경
+                hash.put("ready", (String) payload.get("ready"));
+            }
+            super.setPlayerInfo(roomId, playerId, hash);
 
             /*** Response DTO Build ***/
-            PlayerResponseDto playerResponseDto = PlayerResponseDto.builder()
+            return PlayerResponseDto.builder()
                     .id(playerId)
-                    .nickname((String) super.hashOperations.get(super.getPlayerListKey(roomId, playerId), "nickname"))
-                    .head(Boolean.parseBoolean((String) super.hashOperations.get(super.getPlayerListKey(roomId, playerId), "head")))
-                    .ready(Boolean.parseBoolean((String) super.hashOperations.get(super.getPlayerListKey(roomId, playerId), "ready")))
+                    .ready((boolean) super.getPlayerInfo(roomId, playerId, "ready"))
                     .build();
-
-
-            return playerResponseDto;
         } catch(Exception e) {
             throw new PlayerUpdateException();
         }
@@ -221,14 +217,24 @@ public class PlayerService extends RedisService {
 
     public PlayerResponseDto deletePlayer(Map<String, Object> payload, String roomId) throws PlayerDeleteException {
         try {
+            // 플레이어 존재 확인
+            if(!payload.containsKey("playerId")) {
+                throw new PlayerDeleteException();
+            }
             String playerId = (String) payload.get("id");
-            super.hashOperations.delete(getPlayerListKey(roomId, playerId), "nickname", "head", "ready");
-            playerRepository.deleteById(playerId);
+            if (super.getPlayer(roomId, playerId) == null) {
+                // Redis에도 DB에도 없을 시 Exception
+                if(!playerRepository.existsById(playerId)) {
+                    throw new PlayerDeleteException();
+                }
+            }
 
-            PlayerResponseDto playerResponseDto = PlayerResponseDto.builder()
+            super.deletePlayer(roomId, playerId);  // Redis 삭제
+            playerRepository.deleteById(playerId);  // DB 삭제
+
+            return PlayerResponseDto.builder()
                     .id(playerId)
                     .build();
-            return playerResponseDto;
         } catch(Exception e) {
             throw new PlayerDeleteException();
         }
